Move components into component directory
