+++ web_client/app/components/icons/checkmark.tsx
