+++ web_client/app/pages/registration-check/registration-check.tsx
@@ -3,15 +3,15 @@ import { observer } from 'mobx-react';
 import { Voter } from 'models';
 import { go } from 'router';
 
-import { Link } from 'link/link';
-import { MainContentWrapper } from 'main-content-wrapper/main-content-wrapper';
-import { ShortInput } from 'forms/short-input/short-input';
-import { Button } from 'button/button';
-import { Labelled } from 'forms/labelled/labelled';
+import { Link } from 'components/link/link';
+import { MainContentWrapper } from 'components/main-content-wrapper/main-content-wrapper';
+import { ShortInput } from 'components/forms/short-input/short-input';
+import { Button } from 'components/button/button';
+import { Labelled } from 'components/forms/labelled/labelled';
+import { BirthdayInput } from 'components/forms/birthday-input/birthday-input';
 
 // CSS
 import { styles, vars, css, centeredBox } from 'styles/css';
-import { BirthdayInput } from 'forms/birthday-input/birthday-input';
 
 
 export type RegistrationCheckProps = {
